// src/common/types/url-types.ts
enum URL {
  LINKEDIN = 'https://www.linkedin.com/in/guillermojcrodriguez',
  YOUTUBE = 'https://www.youtube.com/channel/UCluRhwjXllaOlrnY2BKjX_w',
  GITHUB = 'https://github.com/guillermopower',
  BDTGLOBAL = 'https://www.bdtglobal.com/',
  PARAMO = 'https://paramo.tech/',
  SIBLINGS_SOFTWARE = 'https://siblingssoftware.com.ar/',
  TELECENTRO = 'https://telecentro.com.ar/',
  EPAM = 'https://epamneoris.com/',
  CARTASUR = 'https://www.cartasur.com.ar/',
  GARFINAGRO = 'https://www.garfinagro.com/',
  MEDCENTER_SOLUTIONS = 'https://www.linkedin.com/company/medcenter-solutions/',
  UPS = 'https://www.ups.com/ar/es/home',
  NEORIS = 'https://epamneoris.com/',

  MAIL = 'guillermor1977@hotmail.com',
  TYPESCRIPT_BOLIERPLATE = 'https://github.com/manuelbasanta/react-typescript-bolierplate',
  CATALOGO = 'https://vimeo.com/317810816',
  HVS = 'https://github.com/manuelbasanta/HumanVotingSystem',
  ARQCALC = 'https://github.com/guillermopower/arqcalc-ui-ionic',
  PORTFOLIO = 'https://github.com/guillermopower/portfolio',
  BLOCKCHAIN_DEALS = 'https://blockchaindeals.vercel.app/',
}

export default URL;